requirements.txt
streamlit
pandas
segno
fpdf2